Use x-forwarded-for header in ipinfo.io request
If an x-forwarded-for header is among the request's headers, use it in
the query to ipinfo.

This allows this package to be proxied behind nginx or some similar
server and still get the correct result from ipinfo.io.

The proxy server must add x-forwarded-for and set it to the IP address
of the client.

If there's no x-forwarded-for header, no IP address is included as a
part of the query to ipinfo.io; ipinfo.io will determine the IP address
automatically based on source IP address in the IP packet. This means
that you can run this package and its client behind a gateway and still
get meaningful results from ipinfo.io because ipinfo.io is returning
results based on the gateway's public IP address.

Being able to run this package behind a gateway is useful for
development purposes.